refactors up/down logic into separate module
