+++ backend/internal/cmds/mgmt/migrate/up/up.go
@@ -6,14 +6,9 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/golang-migrate/migrate/v4"
-	_ "github.com/golang-migrate/migrate/v4/database/postgres"
-	_ "github.com/golang-migrate/migrate/v4/source/file"
 
 	"github.com/nucleuscloud/neosync/backend/internal/neosyncdb"
+	neomigrate "github.com/nucleuscloud/neosync/internal/migrate"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -58,70 +53,13 @@ func NewCmd() *cobra.Command {
 	return cmd
 }
 
-func getSourceUrl(schemaDir string) (string, error) {
-	var absSchemaDir string
-	if filepath.IsAbs(schemaDir) {
-		absSchemaDir = schemaDir
-	} else {
-		a, err := filepath.Abs(schemaDir)
-		if err != nil {
-			return "", err
-		}
-		absSchemaDir = a
-	}
-
-	return fmt.Sprintf("file://%s", strings.TrimPrefix(absSchemaDir, "file://")), nil
-}
-
-type migrateLogger struct {
-	logger  *slog.Logger
-	verbose bool
-}
-
-func newMigrateLogger(logger *slog.Logger, verbose bool) *migrateLogger {
-	return &migrateLogger{logger: logger, verbose: verbose}
-}
-
-func (m *migrateLogger) Verbose() bool {
-	return m.verbose
-}
-func (m *migrateLogger) Printf(format string, v ...any) {
-	m.logger.Info(fmt.Sprintf("migrate: %s", fmt.Sprintf(format, v...)))
-}
-
 func Up(
 	ctx context.Context,
 	connStr string,
 	schemaDir string,
 	logger *slog.Logger,
 ) error {
-	sourceUrl, err := getSourceUrl(schemaDir)
-	if err != nil {
-		return err
-	}
-
-	m, err := migrate.New(
-		sourceUrl,
-		connStr,
-	)
-	if err != nil {
-		return err
-	}
-	m.Log = newMigrateLogger(logger, false)
-	err = m.Up()
-	if err != nil && !errors.Is(err, migrate.ErrNoChange) {
-		return err
-	}
-
-	serr, derr := m.Close()
-	if serr != nil {
-		return serr
-	}
-	if derr != nil {
-		return derr
-	}
-
-	return nil
+	return neomigrate.Up(ctx, connStr, schemaDir, logger)
 }
 
 func getDbUrl() (string, error) {
