+++ internal/migrate/up.go
@@ -1,0 +1,42 @@
+package neomigrate
+
+import (
+	"context"
+	"errors"
+	"log/slog"
+
+	"github.com/golang-migrate/migrate/v4"
+
+	_ "github.com/golang-migrate/migrate/v4/database/postgres"
+	_ "github.com/golang-migrate/migrate/v4/source/file"
+)
+
+func Up(
+	ctx context.Context,
+	connectionString string,
+	migrationDirectory string,
+	logger *slog.Logger,
+) error {
+	sourceurl, err := getMigrationSourceUrl(migrationDirectory)
+	if err != nil {
+		return err
+	}
+
+	m, err := migrate.New(sourceurl, connectionString)
+	if err != nil {
+		return err
+	}
+	m.Log = newMigrateLogger(logger, false)
+	err = m.Up()
+	if err != nil && !errors.Is(err, migrate.ErrNoChange) {
+		return err
+	}
+	sourcerr, dberr := m.Close()
+	if sourcerr != nil {
+		return sourcerr
+	}
+	if dberr != nil {
+		return dberr
+	}
+	return nil
+}
